PR Feedback: Clean up tests
* Move to separate file, memtests are surprisingly fast
